requirements_simple.txt
gspread==5.12.0
google-auth-oauthlib==1.0.0
requests==2.31.0
python-dotenv==1.0.0
